+++ .circleci/workflows/openlineage-java.yml
@@ -2,6 +2,12 @@ workflows:
   openlineage-java:
     jobs:
       - build-client-java
+      - compile-integration-sql-java-linux
+      - compile-integration-sql-java-macos
+      - build-integration-sql-java:
+          requires:
+            - compile-integration-sql-java-linux
+            - compile-integration-sql-java-macos
       - build-integration-flink:
           matrix:
             parameters:
