+++ integration/sql/iface-java/Cargo.toml
@@ -1,7 +1,7 @@
 [package]
 name = "openlineage_sql_java"
 description = "Java interface for the Rust OpenLineage lineage extraction library"
-version = "0.15.0"
+version = "0.16.0"
 edition = "2021"
 authors = ["Piotr Wojtczak <piotr.m.wojtczak@gmail.com>"]
 keywords = ["sql", "lineage", "openlineage", "java", "jni"]
