+++ integration/sql/iface-java/gradle.properties
@@ -1,0 +1,1 @@
+version=0.16.0-SNAPSHOT
