+++ integration/sql/iface-java/script/build.sh
@@ -1,5 +1,7 @@
 #!/bin/bash
 
+set -e
+
 RED='\033[0;31m'
 GREEN='\033[0;32m'
 NC='\033[0m'
@@ -8,35 +10,26 @@ BASEDIR=$(dirname $BASH_SOURCE)
 ROOT=$BASEDIR/..
 SRC=$ROOT/src
 JAVA_SRC=$SRC/java/io/openlineage/sql
-RESOURCES=$ROOT/resources/io/openlineage/sql
-SCRIPTS=$ROOT/scripts
+RESOURCES=$ROOT/src/main/resources/io/openlineage/sql
+SCRIPTS=$ROOT/script
 
-if [[ "$OSTYPE" == "linux-gnu"* ]]; then
-    NATIVE_LIB_NAME=libopenlineage_sql_java.so
-elif [[ "$OSTYPE" == "darwin"* ]]; then
-    NATIVE_LIB_NAME=libopenlineage_sql_java.dylib
-else
-    printf "\n${RED}Unsupported OS!\n${NC}"
+mkdir -p $RESOURCES
+
+# Native lib should be compiled at this point by compile.sh
+if [[ -f "$ROOT/../target/debug/libopenlineage_sql_java.so" ]]; then
+    cp $ROOT/../target/debug/libopenlineage_sql_java.so $RESOURCES
 fi
-
-# Build the Rust bindings
-cd $ROOT/..
-cargo build -p openlineage_sql_java
-
-cd iface-java
+if [[ "$ROOT/../target/debug/libopenlineage_sql_java.dylib" ]]; then
+    cp $ROOT/../target/debug/libopenlineage_sql_java.dylib $RESOURCES
+fi
 
 # Let's generate this header every run so that it is always
 # up to date.
-mvn dependency:copy-dependencies
+./gradlew clean getDependencies
 ./$SCRIPTS/generate_jni_header.sh
 
 # Package into jar
-mkdir -p $RESOURCES
-cp target/debug/$NATIVE_LIB_NAME $RESOURCES
-mvn assembly:assembly -Dmaven.test.skip
-
-# Run tests
-mvn test
+./gradlew shadowJar
 
 # Run a simple integration test
 printf "\n------ Running integration tests ------\n"
@@ -46,4 +39,5 @@ if [ "$OUTPUT" = "$EXPECTED" ]; then
     printf "\n${GREEN}Integration Tests Passed!${NC}\n"
 else
     printf "\n${RED}Integration Tests Failed!${NC}\n"
+    exit 1
 fi
