+++ integration/sql/iface-java/settings.gradle
@@ -1,0 +1,7 @@
+/*
+ * This file was generated by the Gradle 'init' task.
+ *
+ * This project uses @Incubating APIs which are subject to change.
+ */
+
+rootProject.name = 'openlineage-sql-java'
