+++ integration/sql/iface-java/tests/integration/TestParser.java
@@ -1,3 +1,8 @@
+/*
+/* Copyright 2018-2022 contributors to the OpenLineage project
+/* SPDX-License-Identifier: Apache-2.0
+*/
+
 import io.openlineage.sql.OpenLineageSql;
 import io.openlineage.sql.SqlMeta;
 import java.io.BufferedReader;
@@ -7,7 +12,7 @@ import java.util.Arrays;
 public class TestParser {
     public static void main(String[] args) throws Exception {
         if (args.length > 0) {
-            SqlMeta output = OpenLineageSql.parse(Arrays.asList(args));
+            SqlMeta output = OpenLineageSql.parse(Arrays.asList(args)).get();
             System.out.println(output);
             return;
         }
@@ -16,7 +21,7 @@ public class TestParser {
         BufferedReader reader = new BufferedReader(new InputStreamReader(System.in));
         while (true) {
             String sql = reader.readLine();
-            SqlMeta output = OpenLineageSql.parse(Arrays.asList(sql));
+            SqlMeta output = OpenLineageSql.parse(Arrays.asList(sql)).get();
             System.out.println(output);
         }
     }
