+++ integration/sql/iface-java/tests/java/io/openlineage/sql/OpenLineageSqlTest.java
@@ -1,3 +1,8 @@
+/*
+/* Copyright 2018-2022 contributors to the OpenLineage project
+/* SPDX-License-Identifier: Apache-2.0
+*/
+
 package io.openlineage.sql;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
@@ -10,7 +15,7 @@ import java.util.Arrays;
 class OpenLineageSqlTest {
     @Test
     void basicParse() {
-        SqlMeta output = OpenLineageSql.parse(Arrays.asList("SELECT * FROM test"));
+        SqlMeta output = OpenLineageSql.parse(Arrays.asList("SELECT * FROM test")).get();
         assertEquals(output, new SqlMeta(
             Arrays.asList(new DbTableMeta(null, null, "test")),
             new ArrayList<DbTableMeta>()
@@ -19,7 +24,7 @@ class OpenLineageSqlTest {
 
     @Test
     void parseWithDialect() {
-        SqlMeta output = OpenLineageSql.parse(Arrays.asList("SELECT * FROM `random-project`.`dbt_test1`.`source_table` WHERE id = 1"), "bigquery");
+        SqlMeta output = OpenLineageSql.parse(Arrays.asList("SELECT * FROM `random-project`.`dbt_test1`.`source_table` WHERE id = 1"), "bigquery").get();
         assertEquals(output, new SqlMeta(
             Arrays.asList(new DbTableMeta("random-project", "dbt_test1", "source_table")),
             new ArrayList<DbTableMeta>()
@@ -30,7 +35,7 @@ class OpenLineageSqlTest {
     void parseError() {
         boolean exceptionCaught = false;
         try {
-            SqlMeta output = OpenLineageSql.parse(Arrays.asList("Definitely not an SQL statement"));
+            SqlMeta output = OpenLineageSql.parse(Arrays.asList("Definitely not an SQL statement")).get();
         } catch (Exception e) {
             exceptionCaught = true;
             assertTrue(e instanceof RuntimeException);
